Move `despawn_all` into `World`
